restored the opentelemetry-proto submodule, updated the xtask to use the proto files from the opentelemetry-proto submodule. Renamed grpc_stubs -> proto
