--- rust/otap-dataflow/crates/otlp/proto/opentelemetry/proto/collector/README.md
@@ -1,10 +1,0 @@
-# OpenTelemetry Collector Proto
-
-This package describes the OpenTelemetry collector protocol.
-
-## Packages
-
-1. `common` package contains the common messages shared between different services.
-2. `trace` package contains the Trace Service protos.
-3. `metrics` package contains the Metrics Service protos.
-4. `logs` package contains the Logs Service protos.
