--- rust/otap-dataflow/crates/otlp/proto/opentelemetry/proto/collector/profiles/v1development/profiles_service_http.yaml
@@ -1,9 +1,0 @@
-# This is an API configuration to generate an HTTP/JSON -> gRPC gateway for the
-# OpenTelemetry service using github.com/grpc-ecosystem/grpc-gateway.
-type: google.api.Service
-config_version: 3
-http:
- rules:
- - selector: opentelemetry.proto.collector.profiles.v1development.ProfilesService.Export
-   post: /v1development/profiles
-   body: "*"
