--- rust/otap-dataflow/crates/otlp/proto/opentelemetry/proto/resource/v1/resource.proto
@@ -1,37 +1,0 @@
-// Copyright 2019, OpenTelemetry Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-syntax = "proto3";
-
-package opentelemetry.proto.resource.v1;
-
-import "opentelemetry/proto/common/v1/common.proto";
-
-option csharp_namespace = "OpenTelemetry.Proto.Resource.V1";
-option java_multiple_files = true;
-option java_package = "io.opentelemetry.proto.resource.v1";
-option java_outer_classname = "ResourceProto";
-option go_package = "go.opentelemetry.io/proto/otlp/resource/v1";
-
-// Resource information.
-message Resource {
-  // Set of attributes that describe the resource.
-  // Attribute keys MUST be unique (it is not allowed to have more than one
-  // attribute with the same key).
-  repeated opentelemetry.proto.common.v1.KeyValue attributes = 1;
-
-  // dropped_attributes_count is the number of dropped attributes. If the value is 0, then
-  // no attributes were dropped.
-  uint32 dropped_attributes_count = 2;
-}
