+++ rust/otap-dataflow/crates/otlp/src/grpc.rs
@@ -1,6 +1,6 @@
-use crate::grpc_stubs::proto::collector::logs::v1::ExportLogsServiceRequest;
-use crate::grpc_stubs::proto::collector::metrics::v1::ExportMetricsServiceRequest;
-use crate::grpc_stubs::proto::collector::trace::v1::ExportTraceServiceRequest;
+use crate::proto::opentelemetry::collector::logs::v1::ExportLogsServiceRequest;
+use crate::proto::opentelemetry::collector::metrics::v1::ExportMetricsServiceRequest;
+use crate::proto::opentelemetry::collector::trace::v1::ExportTraceServiceRequest;
 
 /// Expose the OTLP gRPC services.
