+++ rust/otap-dataflow/crates/otlp/src/proto/opentelemetry.proto.collector.logs.v1.rs
