+++ rust/otap-dataflow/crates/otlp/src/proto/opentelemetry.proto.collector.trace.v1.rs
