+++ rust/otap-dataflow/crates/otlp/src/proto/opentelemetry.proto.logs.v1.rs
