+++ rust/otap-dataflow/crates/otlp/src/proto/opentelemetry.proto.metrics.v1.rs
@@ -172,7 +172,7 @@ pub struct Metric {
     #[prost(string, tag = "2")]
     pub description: ::prost::alloc::string::String,
     /// unit in which the metric value is reported. Follows the format
-    /// described by <https://unitsofmeasure.org/ucum.html.>
+    /// described by <http://unitsofmeasure.org/ucum.html.>
     #[prost(string, tag = "3")]
     pub unit: ::prost::alloc::string::String,
     /// Additional metadata attributes that describe the metric. \[Optional\].
@@ -261,7 +261,7 @@ pub struct ExponentialHistogram {
 }
 /// Summary metric data are used to convey quantile summaries,
 /// a Prometheus (see: <https://prometheus.io/docs/concepts/metric_types/#summary>)
-/// and OpenMetrics (see: <https://github.com/prometheus/OpenMetrics/blob/4dbf6075567ab43296eed941037c12951faafb92/protos/prometheus.proto#L45>)
+/// and OpenMetrics (see: <https://github.com/OpenObservability/OpenMetrics/blob/4dbf6075567ab43296eed941037c12951faafb92/protos/prometheus.proto#L45>)
 /// data type. These data points cannot always be merged in a meaningful way.
 /// While they can be useful in some applications, histogram data points are
 /// recommended for new applications.
@@ -373,9 +373,7 @@ pub struct HistogramDataPoint {
     /// The sum of the bucket_counts must equal the value in the count field.
     ///
     /// The number of elements in bucket_counts array must be by one greater than
-    /// the number of elements in explicit_bounds array. The exception to this rule
-    /// is when the length of bucket_counts is 0, then the length of explicit_bounds
-    /// must also be 0.
+    /// the number of elements in explicit_bounds array.
     #[prost(fixed64, repeated, tag = "6")]
     pub bucket_counts: ::prost::alloc::vec::Vec<u64>,
     /// explicit_bounds specifies buckets with explicitly defined bounds for values.
@@ -391,9 +389,6 @@ pub struct HistogramDataPoint {
     /// Histogram buckets are inclusive of their upper boundary, except the last
     /// bucket where the boundary is at infinity. This format is intentionally
     /// compatible with the OpenMetrics histogram definition.
-    ///
-    /// If bucket_counts length is 0 then explicit_bounds length must also be 0,
-    /// otherwise the data point is invalid.
     #[prost(double, repeated, tag = "7")]
     pub explicit_bounds: ::prost::alloc::vec::Vec<f64>,
     /// (Optional) List of exemplars collected from
