+++ rust/otap-dataflow/crates/otlp/src/proto/opentelemetry.proto.resource.v1.rs
