+++ rust/otap-dataflow/xtask/src/main.rs
@@ -80,15 +80,21 @@ fn test_all() -> anyhow::Result<()> {
 
 fn compile_proto_otlp() -> anyhow::Result<()> {
     tonic_build::configure()
-        .out_dir("crates/otlp/src/grpc_stubs")
+        .out_dir("crates/otlp/src/proto")
         .compile_protos(
             &[
-                "crates/otlp/proto/opentelemetry/proto/collector/logs/v1/logs_service.proto",
-                "crates/otlp/proto/opentelemetry/proto/collector/metrics/v1/metrics_service.proto",
-                "crates/otlp/proto/opentelemetry/proto/collector/trace/v1/trace_service.proto",
+                "opentelemetry/proto/common/v1/common.proto",
+                "opentelemetry/proto/resource/v1/resource.proto",
+                "opentelemetry/proto/trace/v1/trace.proto",
+                "opentelemery/proto/metrics/v1/metrics.proto",
+                "opentelemery/proto/logs/v1/logs.proto",
+                "opentelemery/proto/collector/logs/v1/logs_service.proto",
+                "opentelemery/proto/collector/trace/v1/trace_service.proto",
+                "opentelemetry/proto/collector/metrics/v1/metrics_service.proto",
             ],
-            &["crates/otlp/proto"],
-        )?;
+            &["../../proto/opentelemetry-proto"],
+        )
+        .expect("Failed to compile OTLP protos.");
     Ok(())
 }
